+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2.0.9
+
+### Fixed
+
+- Fix to ensure updateJWT works when changing currency from original JWT
+- Update StCodec to always publish to parent
+- Update ApplePay successCallback to be called on success
+
 ## 2.0.8
 
 ### Fixed